backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
bcrypt==4.1.2
anthropic==0.18.1
google-generativeai==0.8.0
groq==0.4.0
openai==1.12.0
openpyxl==3.1.2
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
